add gather_elements operation
GatherElements takes two inputs data and indices of the same rank r >= 1
and an attribute axis that identifies an axis of data.
 It is an indexing operation that produces its output by indexing into
 the input data tensor at index positions determined by elements of
the indices tensor. Its output shape is the same as the shape of indices and
consists of one value (gathered from the data) for each element in indices.

update ovxlib version to 1.1.41

Issue:31530

Change-Id: I6786c36bf5d6d11cf17c70b2e316f06c7bcbc869