add scatter_elements operations
ScatterElements takes three inputs data, updates, and indices of
the same rank r >= 1 and an optional attribute axis that identifies
an axis of data (by default, the outer-most axis, that is axis 0).
The output of the operation is produced by creating a copy of the input data,
and then updating its value to values specified by updates at specific
index positions specified by indices. Its output shape is the same as the shape of data

Change-Id: I7c24c33329baa534c636ad45dc6d3992bdc49608