add softmax z direction sp kernel
This program requires MAX/EXP operation in the Y dimension.
The Element Marker bits are not useful here because we are performing
the Max in the Y dimension not in the X dimension.
 As a result, this program has a limitation that HWTileXSize = ImageXSize.
If ImageXSize is too large, software is responsible to split the original
 image into columns of smaller images.

fix layernorm sp kernel instruction issue
fix lstmunit activation sp kernel attribute issue

Change-Id: I4a8cd1a6475b29980774e76702ef978e20be61e1