Use vxRemoveKernel to remove the user kernels
For user kernels created with vxAddUserKernels or vxImportKernelFromURL,
we should use vxRemoveKernel to release the related resource and remove
them from kernel table when they are not used any more.

When user kernel is created, an unsued kernel object is allocated from
kernel table for it. And the internal reference count as well as external
reference count are increased to the kernel.

vxRemoveKernel will decrease the internal and external reference count to
release the resource and remove it from the kernel table.

vxReleaseKernel just decreases the external reference count. The resource
is not released and the kernel is not removed from the kernel table as the
non-zero internal reference count.

Issue: ARTPEC8-361
Change-Id: I376db137e52eb08589c33a337758a098920e9d6d
Type: Bug Fix