add cache for _check_swapped_tensors
_check_swapped_tensors will traverse all input & output tensors
of NBG node.
It will be very slow, when NBG node has many input & output
tensors.
If you always swap handle of the same tensors in the same graph
during the different running of the graph, you can use this
feature to cache the information of swap handle in the first
time you run the graph.
You can use vsi_nn_SwapTensorHandleWithCache to switch on this
feature.

Change-Id: I820bb3ddd094efa8f513cb25d2c41da6eea015e3
Type: New Feature