Extend the resize op to support cubic feature
Calculate the x_int, y_int and 4 interplation coefficients corresponding
to each point in a row and column based on the width and height of the
output. The shader kernel reads the x_int, y_int and their respective
coefficients at the output coordinates to perform interpolation and obtain
the final result.

Change-Id: I7ea2fc03d74b8e9295eb2dd1f0ee2dc9c70f7717
Type: New Feature
Issue: acuity 1261
Signed-off-by: Haoxiang.Wang <Haoxiang.Wang@verisilicon.com>